Adds Status repo and moves mastodon4J extensions to base
