+++ app/src/main/java/io/github/koss/mammut/extension/LiveDataExtensions.kt
@@ -1,10 +1,7 @@
 package io.github.koss.mammut.extension
 
 import androidx.lifecycle.*
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlinx.coroutines.*
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
@@ -19,28 +16,30 @@ fun <T> LiveData<T>.postSafely(item: T) {
 fun <X> LiveData<List<X>>.filterElements(predicate: (X) -> Boolean): LiveData<List<X>> {
     return MediatorLiveData<List<X>>().apply {
         addSource(this@filterElements) { x ->
-           value = x.filter(predicate)
+            value = x.filter(predicate)
         }
     }
 }
 
-suspend fun <T> LiveData<T>.awaitFirst(): T = suspendCancellableCoroutine {
-    var resumed = false
-    // Try short circuit
-    value?.let { value ->
-        it.resume(value)
-        resumed = true
-        return@suspendCancellableCoroutine
-    }
-
-    // Wait for first emission
-    val observer = { value: T ->
-        if (!resumed) {
+suspend fun <T> LiveData<T>.awaitFirst(): T = coroutineScope {
+    suspendCancellableCoroutine<T> {
+        var resumed = false
+        // Try short circuit
+        value?.let { value ->
             it.resume(value)
             resumed = true
+            return@suspendCancellableCoroutine
         }
-    }
-    GlobalScope.launch (Dispatchers.Main) {
-        observeForever(observer)
+
+        // Wait for first emission
+        val observer = { value: T ->
+            if (!resumed) {
+                it.resume(value)
+                resumed = true
+            }
+        }
+        launch(Dispatchers.Main) {
+            observeForever(observer)
+        }
     }
 }
