+++ app/src/main/java/io/github/koss/mammut/feature/instance/subfeature/profile/ProfileViewModel.kt
@@ -14,7 +14,7 @@ import io.github.koss.mammut.data.database.MammutDatabase
 import io.github.koss.mammut.data.models.Account
 import io.github.koss.mammut.data.models.NetworkState
 import io.github.koss.mammut.extension.postSafely
-import io.github.koss.mammut.extension.run
+import io.github.koss.mammut.base.extensions.run
 import io.github.koss.mammut.feature.instance.dagger.InstanceScope
 import io.github.koss.mammut.feature.instance.subfeature.profile.dagger.ProfileScope
 import kotlinx.coroutines.CoroutineScope
