+++ base/src/main/java/io/github/koss/mammut/base/extensions/Mastodon4JExtensions.kt
@@ -1,0 +1,58 @@
+package io.github.koss.mammut.base.extensions
+
+import android.util.Log
+import arrow.core.Either
+import arrow.core.Left
+import arrow.core.Right
+import com.google.gson.GsonBuilder
+import com.google.gson.annotations.Expose
+import com.google.gson.annotations.SerializedName
+import com.sys1yagi.mastodon4j.MastodonRequest
+import com.sys1yagi.mastodon4j.api.exception.Mastodon4jRequestException
+import io.github.koss.mammut.base.BuildConfig
+
+
+typealias MastodonResult<T> = Either<Error, T>
+
+tailrec suspend fun <T> MastodonRequest<T>.run(retryCount: Int = 0): Either<Error, T> {
+    val result = try {
+        Right(execute())
+    } catch (e: Mastodon4jRequestException) {
+        Log.e("MastodonRunner", "An error occurred", e)
+        if (e.isErrorResponse()) {
+            e.response?.body()?.string()?.run {
+                when {
+                    startsWith("{") ->  try {
+                        Left(GsonBuilder()
+                                .excludeFieldsWithoutExposeAnnotation()
+                                .create()
+                                .fromJson<Error>(e.response?.body()?.charStream(), Error::class.java))
+                    } catch (e: Exception) {
+                        null
+                    }
+                    isNotEmpty() -> unknownError(this)
+                    else -> null
+                }
+            } ?: unknownError(if (BuildConfig.DEBUG) "Exception from non-error response" else "Oh jeez, something's gone wrong")
+        } else {
+            unknownError(if (BuildConfig.DEBUG) "Exception from non-error response" else "Oh jeez, something's gone wrong")
+        }
+    }
+
+    if (result is Either.Left && retryCount > 0) {
+        return run(retryCount - 1)
+    }
+
+    return result
+}
+
+private fun unknownError(description: String) =
+        Left(Error(error = "unknown_error", description = description))
+
+data class Error(
+        @Expose
+        val error: String,
+        @Expose
+        @SerializedName("error_description")
+        val description: String
+)
