+++ build.gradle
@@ -62,7 +62,7 @@ ext {
     calligraphy = "uk.co.chrisjenx:calligraphy:$calligraphy_version"
 
     // Mastodon4J
-    mastodon4j = 'com.github.jamiesanson.mastodon4j:mastodon4j:e410f1c137'
+    mastodon4j = 'com.github.jamiesanson.mastodon4j:mastodon4j:6271fb978c'
 
     // Room
     def room_version = "1.1.1"
